Reorganized ReportFilterControls and its context
